ENH: Import of new work on 2D API and charts.
This is the initial import of the work I have been doing on the 2D API
and charts in VTK. Currently there is a fairly well developed OpenGL
backend (with limited primitives) and a very early Qt Painter backend
that is more a proof of concept (many features do not work).

Please see the wiki pages on the new 2D API and charts in VTK.

http://www.vtk.org/Wiki/VTK/Charts

Please address any initial feedback to the InfoVis mailing list.