ENH: Prototype of a scene/canvas API for VTK.
The context scene manages the context items. The block item is an
example implementation of an interactive scene item. The context actor
is essentially the view - provides an abject to add to a VTK window to
be rendered. All rendering is done in the overlay using the 2D context
API.