Only process files with vtkCxxDescriptorMacro present.
This should un-break the branch.

Also, fix descriptor parser to handle member specifications with
no whitespace between the final argument and the terminating semicolon;
the last argument was being omitted in this case.

Change the docs and code so that, instead of requiring a mark on
RW-access members to be included in a serialization, you must now
add a mark to those members you *don't* want serialized. This should
require less typing since pretty much everything read-write should be
saved.